+++ lib/render/index.js
@@ -1,6 +1,7 @@
 const path = require('path')
 const Renderer = require('./renderer')
 const Source = require('./source')
+const Rss = require('./rss')
 const Helper = require('./helper')
 const config = require('../config')
 
@@ -10,6 +11,7 @@ class Render extends Renderer {
     this.config = config
     this.helper = new Helper()
     this.source = new Source()
+    this.rss = new Rss()
   }
 
   _path(tag) {
@@ -102,6 +104,7 @@ class Render extends Renderer {
 
     if (!config) {
       this.copySources()
+      this.rss.generate()
       return this.genHtmls(data, 'all')
     }
